♻️ Refactor: store 폴더를 src/로 이동 및 import 경로 수정
